Rearranged Project files, Added 4 http request verbs
The entire project was organised into separate sections; one for Python
and one for Java. The Python side runs the API locally with Flask. The
Java side will be the actual GUI along with the request methods to the
API.

4 HTTP request methods were added; 3 of which work. These three are GET
all longboards, DELETE clear all longboards and OPTIONS for the
/longboards endpoint. The final method, POST new longboard, is not
working in this commit. The 'id' argument must be first in the body, but
for some reason when using put to insert the argument into the HashMap
that is passed into the Unirest POST method, 'id' goes to the back.
Because the 'id' is not found, it returns a 400 error- Bad Request.