Rewrote http Java methods using switch statements
Before, there were many redundant try/catch statements and many of the
requests were extremely similar. As an attempt to revise this, the
methods of high similarity were combined into one method that takes an
addition argument. This argument, requestType, distinguishes which http
verb the request should make. This is done with a switch statement.
Overall, I think it improves the code just a bit.