Variety of changes acroos all files
First, the Main program was rewritten (again) and is now about 40 lines
shorter than previously despite the same functionality. Two new methods
replaced addElements and they are addLabeledUserInput and addJButton.
Both of these will be used even more in the future.

Next, with regards to the files in src, the simple requests in both
specific and general longboard requests had a minor update; if the
default response body is null, return a string instead. This is so that
something helpful is returned to the Main file to be displayed to user
in the popup message.

Lastly, this is a very minor change, but the requestType parameter switch
case in general simple longboard requests "clear" has been changed to
"delete". This was done because the Action Listeners now pass a string
that is given as an argument straight down to the HTTP methods.