Final commit; README.md updated
The readme could definetly have more content and in all honesty it
should have more. BUT I am really excited to move on to my next project
and I am overall happy with all that I learned on this one as well as
the product. Even comparing it to the first major Java project I did
(the password bank) it is almost the same length(the main.java files) yet there is so so much
more functionality in this project which, to me, is a sure sign of
improvement.